+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/update/flush/DirtyAttributeFlusher.java
@@ -65,6 +65,8 @@ public interface DirtyAttributeFlusher<T extends DirtyAttributeFlusher<T, E, V>,
 
     public AttributeAccessor getEntityAttributeAccessor();
 
+    public String getLockOwner();
+
     public boolean isOptimisticLockProtected();
 
     public boolean requiresFlushAfterPersist(V value);