+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/update/subview/inverse/embedded/simple/model/UpdatableLegacyOrderPositionEmbeddableView.java
@@ -16,34 +16,23 @@
 
 package com.blazebit.persistence.view.testsuite.update.subview.inverse.embedded.simple.model;
 
-import com.blazebit.persistence.view.CreatableEntityView;
-import com.blazebit.persistence.view.EntityView;
-import com.blazebit.persistence.view.InverseRemoveStrategy;
-import com.blazebit.persistence.view.MappingInverse;
-import com.blazebit.persistence.view.UpdatableEntityView;
-import com.blazebit.persistence.view.UpdatableMapping;
-import com.blazebit.persistence.view.testsuite.entity.LegacyOrderPosition;
+import java.io.Serializable;
 
-import java.util.Set;
+import com.blazebit.persistence.view.EntityView;
+import com.blazebit.persistence.view.UpdatableEntityView;
+import com.blazebit.persistence.view.testsuite.entity.LegacyOrderPosition;
+import com.blazebit.persistence.view.testsuite.entity.LegacyOrderPositionEmbeddable;
+import com.blazebit.persistence.view.testsuite.entity.LegacyOrderPositionId;
 
 /**
  *
  * @author Christian Beikov
- * @since 1.2.0
+ * @since 1.6.0
  */
-@CreatableEntityView
 @UpdatableEntityView
-@EntityView(LegacyOrderPosition.class)
-public interface UpdatableLegacyOrderPositionView extends LegacyOrderPositionIdView {
+@EntityView(LegacyOrderPositionEmbeddable.class)
+public interface UpdatableLegacyOrderPositionEmbeddableView extends Serializable {
 
-    String getArticleNumber();
-    void setArticleNumber(String articleNumber);
-
-    UpdatableLegacyOrderPositionEmbeddableView getEmbeddable();
-    void setEmbeddable(UpdatableLegacyOrderPositionEmbeddableView embeddable);
-
-    @UpdatableMapping
-    @MappingInverse(removeStrategy = InverseRemoveStrategy.REMOVE)
-    Set<LegacyOrderPositionDefaultIdView> getDefaults();
-    void setDefaults(Set<LegacyOrderPositionDefaultIdView> defaults);
+    String getName();
+    void setName(String name);
 }